Allow to override build date with SOURCE_DATE_EPOCH
in order to make builds reproducible.
See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

Also use UTC to be independent of timezone.
This date call only works with GNU date.

This patch was imported from https://bugs.debian.org/886386
and successfully tested on openSUSE.

Signed-off-by: Bernhard M. Wiedemann <bwiedemann@suse.de>